model_inference/pytorch/requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html

# GPU Accelerated
# torchvision==0.11.1+cu113

# CPU Only
torch==1.10.0+cpu
torchvision==0.11.1+cpu

torchinfo==1.6.3
pytorch-ignite==0.4.7
avro==1.11.0
confluent-kafka==1.8.2
requests==2.27.1
pika==1.3.2